EPMRPP-104683: moved Divider to the common components
