+++ app/src/pages/common/divider/divider.scss
