+++ app/src/pages/inside/projectSettingsPageContainer/content/elements/index.js
@@ -17,7 +17,6 @@
 export { FieldElement } from './fieldElement';
 export { LabeledPreloader } from './labeledPreloader';
 export { TabDescription } from './tabDescription';
-export { Divider } from './divider';
 export { RuleList, DraggableRuleList } from './ruleList';
 export { PatternRuleContent } from './patternRuleContent';
 export { NotificationRuleContent } from './notificationRuleContent';
